+++ changelogs/unreleased/8755-sseago
@@ -1,0 +1,1 @@
+Move pvc annotation removal from CSI RIA to regular PVC RIA
